requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
supabase>=1.0.0
apscheduler>=3.10.0
pytz>=2023.3
python-dateutil==2.8.2 